api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
psycopg[binary]==3.1.19
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
boto3==1.34.162
python-multipart==0.0.9
PyYAML==6.0.1
